refactor(session-replay): Add injectable mask and view renderer (#4938)
