+++ Sources/Sentry/SentryScreenshot.m
@@ -16,7 +16,8 @@
 {
     if (self = [super init]) {
         photographer = [[SentryViewPhotographer alloc]
-            initWithRedactOptions:[[SentryRedactDefaultOptions alloc] init]];
+            initWithRenderer:[[SentryDefaultViewRenderer alloc] init]
+               redactOptions:[[SentryRedactDefaultOptions alloc] init]];
     }
     return self;
 }
