+++ Sources/Swift/Tools/ViewCapture/SentryDefaultViewRenderer.swift
@@ -1,0 +1,17 @@
+#if canImport(UIKit) && !SENTRY_NO_UIKIT
+#if os(iOS) || os(tvOS)
+
+import UIKit
+
+@objcMembers
+class SentryDefaultViewRenderer: NSObject, SentryViewRenderer {
+    func render(view: UIView) -> UIImage {
+        let image = UIGraphicsImageRenderer(size: view.bounds.size).image { _ in
+            view.drawHierarchy(in: view.bounds, afterScreenUpdates: false)
+        }
+        return image
+    }
+}
+
+#endif // os(iOS) || os(tvOS)
+#endif // canImport(UIKit) && !SENTRY_NO_UIKIT
