+++ Sources/Swift/Tools/ViewCapture/SentryViewRenderer.swift
@@ -1,0 +1,11 @@
+#if canImport(UIKit) && !SENTRY_NO_UIKIT
+#if os(iOS) || os(tvOS)
+
+import UIKit
+
+@objc protocol SentryViewRenderer {
+    func render(view: UIView) -> UIImage
+}
+
+#endif // os(iOS) || os(tvOS)
+#endif // canImport(UIKit) && !SENTRY_NO_UIKIT
