+++ Sources/Swift/Tools/ViewCapture/SentryViewScreenshotProvider.swift
