+++ Sources/Swift/Tools/ViewCapture/UIRedactBuilder.swift
