+++ Tests/SentryTests/ViewCapture/UIRedactBuilderTests.swift
